Changed Navx to pigeon and added more subsystems
Though Navx was going to be used but apparently not. New subsystem is added for the Pigeon gyro along with a few other subsystems(ModuleIOSim and ModuleIOSpark). There still needs to be more integration in the added files to change some of the variables to match the ones in constants. Also looking to see if any of the new subsystems repeat the swerve drive calculations(quite a bit of it looked familiar but may be for any simulations)